✨ Add physical buttons for cat sitters
Up until this point, we've controlled the system
using scripts, but a family member needs to be able
to control things with the push of a button.

The biggest concern is disabling the motion sensors
while someone is resetting the system. If things are pulled
apart and the motion sensors go off, they'll dispense food
and make a big mess.